test(path-layer): colocate antialiasing render coverage
